+++ source/Burntime.Remaster/Logic/Item/ItemTypes.cs
@@ -103,7 +103,7 @@ namespace Burntime.Remaster.Logic
 
         public string[] GetTypesWithClass(string[] include, string[] exclude)
         {
-            List<string> list = new List<string>();
+            var list = new List<string>();
 
             foreach (ItemType type in types)
             {
@@ -111,7 +111,7 @@ namespace Burntime.Remaster.Logic
                 bool excluded = false;
                 foreach (string e in exclude)
                 {
-                    if (type.IsClass(e))
+                    if (type.IsClass(e) || type.ID == e)
                     {
                         excluded = true;
                         break;
@@ -125,7 +125,7 @@ namespace Burntime.Remaster.Logic
                 // now look for included flags
                 foreach (string i in include)
                 {
-                    if (type.IsClass(i))
+                    if (type.IsClass(i) || type.ID == i)
                     {
                         // if included then add to list
                         list.Add(type.ID);